requirements.txt
django>=3.2.8
djangorestframework>=3.12.4
django-rest-passwordreset>=1.2.0
drf-yasg>=1.20.0
